split the helper functions and the sc kernel to two files
