+++ tests/test_parse_ase_otf.py
@@ -1,6 +1,6 @@
 import numpy as np
 from flare.otf_parser import OtfAnalysis
-from flare.kernels.kernels import two_plus_three_body, two_plus_three_body_grad
+from flare.kernels.sc import two_plus_three_body, two_plus_three_body_grad
 from flare.kernels.mc_simple import two_plus_three_body_mc, two_plus_three_body_mc_grad
 from flare.env import AtomicEnvironment
 from flare.predict import predict_on_structure
